+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New features
 
+* New interactive command `cider-undef`.
 * First pass at a CIDER quick reference card.
 
 * `completion-at-point` now annotates functions, macros and special forms, thus making it